\name{get_simulation_results}
\alias{get_simulation_results}
\title{
Retrieve time series of simulated variables
}
\description{
The \code{get_simulation_results} function retrieves the time series of simulated variables
from an object of \code{gecon_simulation} class.
}
\usage{
get_simulation_results(sim_obj)
}
\arguments{
  \item{sim_obj}{
An object of \code{\link{gecon_simulation-class}} class.
}
}
\value{
The results are returned as a matrix if the simulation has been generated by
a call to the \code{random_path} or \code{simulate_model} function or a list of matrices corresponding to
selected shocks in a call to the \code{compute_irf} function.
}
\seealso{
For details, see \code{\link{gecon_simulation-class}}.
}
\examples{
# copy the example to the current working directory
file.copy(from = file.path(system.file("examples", package = "gEcon"),
          "rbc.gcn"), to = getwd())

# make and load the model
rbc <- make_model("rbc.gcn")

# solve the model
rbc <- steady_state(rbc)
rbc <- solve_pert(rbc)

# compute and print the IRFs
rbc <- set_shock_cov_mat(rbc, cov_matrix = matrix(0.01, 1, 1),
                         shock_order = "epsilon_Z")
irf_rbc <- compute_irf(rbc, cholesky = TRUE,
                       variables = c("K_s", "C", "Z", "I", "Y"))
get_simulation_results(irf_rbc)
}
